flask-api/app/requirements.txt
celery
redis
langchain
langchain-community
langchain_openai
openai
PyPDF2
flask
flask-restful
python-dotenv
pypdf
chromadb